in SpdyPersistentUrl, call the retryCallback inline, instead of asynchronously on the main thread as before.
This is because we want to allow suppliers of the retryCallback to immediately start
a new background task (if applicable) so that their app does not get suspended
while running in the background.

In the case that the app will be suspended without a background task, the retryCallback
will not run until the app resumes, which may be never.