update SpdyPersistentUrl to use the new CTTelephonyNetworkInfo radio access technology api in ios7 which lets us know when we have moved from LTE to CDMA, etc.
If we are connected via WWAN, and get a radio access change from our previous state, then we send a ping
to make sure that the connection is still valid.

If it is not, we will then reconnect in short order, instead of waiting for the next 10 minute socket check.